final spritesheet + character descriptions
Zafra
seems like a normal sullen teenager, but quite murderous when it comes down to it!
Mission is to slay the werewolves plaguing the village near the forest. She’s not here to help the villagers. She’s here because killing is fun.
She’s somebody who hasn’t really learned the virtue of patience yet. So hurry up and pack the bag!!

Tiberius 
golfing with the boys has gotten boring. Fishing with the boys has gotten boring too! This guy needs some more excitement in his life!
He’s rather laidback about life. When you get to his age, nothing can really ruin your day!
Mission is to find a cool treasure. something fun for the mantlepiece, to show his grown-up children.
His pride and joy is his luscious hair and beard. God forbid anything happens to it…

Goldie 
here to fetch, fetch, fetch!
Here to get a necklace back from the swamp! Unfortunately, the swamp is full of slime, goblins who have recently learned archery, and all sorts of unpleasant things that would get a dawg down.
But ultimately he’s an optimist at heart :) so he’ll get back up and keep trying!
attempted to enlist in army. Was too much of a good boy and not enough of an attack dog to be let in.